增加mpa 支持多页面打包，完美

+++ scripts/build.js
@@ -16,19 +16,31 @@ module.exports = function() {
 
     preinstall();
 
-    if (fs.existsSync(output)){
-      fse.removeSync(output);
+    if(fs.existsSync(path.join(__dirname, `../packages/${type}/webpack.build.js`))){
+      const webpack = require('webpack');
+      const webpackConfig = require(path.join(
+        __dirname,
+        `../packages/${type}/webpack.build`,
+      ));
+  
+      const compiler = webpack(webpackConfig);
+
+      if (fs.existsSync(output)){
+        log(`webpack remove prevs exist output directory. ${output.red}`);
+        fse.removeSync(output);
+      }
+  
+      compiler.run(() => log('webpack building completed!'));
+    }else{
+      // 自定义发布方法
+      const customBuildServer = require(path.join(
+        __dirname,
+        `../packages/${type}/build`,
+      ));
+
+      return customBuildServer();
     }
 
-    const webpack = require('webpack');
-    const webpackConfig = require(path.join(
-      __dirname,
-      `../packages/${type}/webpack.build`,
-    ));
-
-    const compiler = webpack(webpackConfig);
-
-    compiler.run(() => log('webpack building completed!'));
   } else {
     error(`Can not find ${'abc.json'.bold} in current directory`.red);
   }
